feat/#330: 추천 탭 본앱 연동
- AppCoordinator, AppDelegate에 RecommendationFeature 등록
- MLSCore에 TokenInterceptor 추가 (키체인 직접 접근)
- RecommendationRepositoryImpl 내부에서 NetworkProvider, Interceptor 생성
- UserProfileDTO, JobDTO 응답 래퍼 구조 반영
- RecommendationEndPoint 쿼리 파라미터 Int32 타입 적용
- viewWillAppear마다 데이터 갱신
- 추천 화면 탭바 영역 레이아웃 수정
- MLSCore NetworkProviderImpl 로깅 Loggable 적용

+++ MLS/MLSRecommendationFeature/Sources/MLSRecommendationFeature/Data/DTOs/RecommendationMapDTO.swift
@@ -2,7 +2,7 @@ import MLSRecommendationFeatureInterface
 
 struct RecommendationMapDTO: Decodable {
     let mapId: Int
-    let score: Int
+    let score: Double
     let iconUrl: String
     let nameKr: String
     let bookmarkId: Int?
